refactor: isolate legacy recovery compatibility
